+++ Package.swift
@@ -13,7 +13,7 @@ let package = Package(
         ),
     ],
     dependencies: [
-        .package(url: "https://github.com/patrickridd/AuthDomain.git", from: "1.1.0"),
+        .package(url: "https://github.com/patrickridd/AuthDomain.git", from: "1.2.0"),
         .package(url: "https://github.com/firebase/firebase-ios-sdk.git", from: "12.14.0"),
         .package(url: "https://github.com/google/GoogleSignIn-iOS.git", from: "8.0.0"),
         .package(url: "https://github.com/facebook/facebook-ios-sdk.git", from: "18.0.0"),